+++ source/compress_integer_simple_9.h
@@ -58,9 +58,9 @@ namespace JASS
 			class lookup
 				{
 				public:
-					size_t numbers;							///< Number of integers encoded using this selector.
-					size_t bits;								///< NUmber of vits used to encode each integer.
-					size_t mask;								///< A bitmast to select an integer.
+					uint32_t numbers;							///< Number of integers encoded using this selector.
+					uint32_t bits;								///< NUmber of vits used to encode each integer.
+					uint32_t mask;								///< A bitmast to select an integer.
 				};
 
 		protected:
@@ -68,14 +68,14 @@ namespace JASS
 				internal tables for comptability with derived schemes
 			*/
 			static const lookup simple9_table[];				///< The table mapping bits to slectors and masks
-			static const size_t bits_to_use[];					///< The number of bits used to store an integer of the given the number of bits in length
-			static const size_t table_row[];						///< Given the number of bits, which row of simple9_table should be used?
+			static const uint32_t bits_to_use[];					///< The number of bits used to store an integer of the given the number of bits in length
+			static const uint32_t table_row[];						///< Given the number of bits, which row of simple9_table should be used?
 
-			static const size_t ints_packed_table[];			///< Number of integers packed into a 32-bit word, given its mask type
-			static const size_t can_pack_table[];				///< Bitmask map for valid masks at an offset (column) for some num_bits_needed (row).
-			static const size_t row_for_bits_needed[];		///< Translates the 'bits_needed' to the appropriate 'row' offset for use with can_pack table.
-			static const size_t invalid_masks_for_offset[];	///< AND out masks for offsets where we don't know if we can fully pack for that offset
-			static const size_t simple9_shift_table[];		///< Number of bits to shift when packing - 9 rows for simple-9
+			static const uint32_t ints_packed_table[];			///< Number of integers packed into a 32-bit word, given its mask type
+			static const uint32_t can_pack_table[];				///< Bitmask map for valid masks at an offset (column) for some num_bits_needed (row).
+			static const uint32_t row_for_bits_needed[];		///< Translates the 'bits_needed' to the appropriate 'row' offset for use with can_pack table.
+			static const uint32_t invalid_masks_for_offset[];	///< AND out masks for offsets where we don't know if we can fully pack for that offset
+			static const uint32_t simple9_shift_table[];		///< Number of bits to shift when packing - 9 rows for simple-9
 
 
 		public:
